Fix entity schema
The entity schema is now in correspondence with `session.sql`
(previously `session_install.sql`). There's only one difference: Fields
that have a foreign key with `ON DELETE CASCADE` are required (`NOT
NULL`) now. This change is applied in `upgrade_0017()`.

`session_install.sql` is renamed to `session.sql` to be only executed on
update, not on install.