Allow to specify entity field name in field spec
This change allows to use names for form fields different to the entity
field names. Additionally it's now possible to specify value callbacks
in the field spec (e.g. if the form field value is combined of two
values for the dependent field functionality).

This fixes also loading of country_id and state_province_id (see code
comments for details.)